Use test item framework
