+++ Project.toml
@@ -22,6 +22,7 @@ SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
 LibGit2 = "76f85450-5226-5b5a-8eaa-529ad045b433"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
+TestItemRunner = "f8b46487-2199-4994-9208-9a1283c18c0a"
 
 [compat]
 JSON = "0.20, 0.21"
@@ -35,4 +36,4 @@ JSONRPC = "1.1"
 SymbolServer = "7.1"
 
 [targets]
-test = ["Test", "Sockets", "LibGit2", "Serialization", "SHA"]
+test = ["Test", "Sockets", "LibGit2", "Serialization", "SHA", "TestItemRunner"]
